Fix operational reads after dough production cleanup

+++ src/ParlorPrediction.Contracts/Requests/DoughUsage/CorrectDoughUsageTraceRequest.cs
@@ -10,7 +10,7 @@ public sealed class CorrectDoughUsageTraceRequest
 
     public string Destination { get; init; } = string.Empty;
 
-    public int TrayCount { get; init; }
+    public decimal TrayCount { get; init; }
 
     public string? Notes { get; init; }
